test(java): :white_check_mark: web automation test in progress
