+++ sample-tests/src/test/java/com/github/wasiqb/boyka/api/requests/TokenBuilder.java
@@ -1,0 +1,15 @@
+package com.github.wasiqb.boyka.api.requests;
+
+/**
+ * @author Faisal Khatri
+ * @since 22/07/2022
+ */
+public class TokenBuilder {
+    public Tokencreds tokenBuilder () {
+        return Tokencreds.builder ()
+            .username ("admin")
+            .password ("password123")
+            .build ();
+    }
+
+}
