+++ sample-tests/src/test/java/com/github/wasiqb/boyka/api/requests/Tokencreds.java
@@ -1,0 +1,15 @@
+package com.github.wasiqb.boyka.api.requests;
+
+import lombok.Builder;
+import lombok.Getter;
+
+/**
+ * @author Faisal Khatri
+ * @since 22/07/2022
+ */
+@Getter
+@Builder
+public class Tokencreds {
+    private String password;
+    private String username;
+}
