--- sample-tests/src/test/java/com/github/wasiqb/boyka/apitests/restfulbooker/requests/Tokencreds.java
@@ -1,15 +1,0 @@
-package com.github.wasiqb.boyka.apitests.restfulbooker.requests;
-
-import lombok.Builder;
-import lombok.Getter;
-
-/**
- * @author Faisal Khatri
- * @since 22/07/2022
- */
-@Getter
-@Builder
-public class Tokencreds {
-    private String username;
-    private String password;
-}
