--- sample-tests/src/test/java/com/github/wasiqb/boyka/uitests/testdata/RegisterUserData.java
@@ -1,19 +1,0 @@
-package com.github.wasiqb.boyka.uitests.testdata;
-
-import lombok.Builder;
-import lombok.Getter;
-
-/**
- * @author Faisal Khatri
- * @since 8/3/2022
- **/
-@Builder
-@Getter
-public class RegisterUserData {
-
-    private String email;
-    private String firstName;
-    private String lastName;
-    private String password;
-    private String telephone;
-}
